refactor(tocco-ui): fix Upload.js
Refs: TOCDEV-4547
Changelog: fix UploadInput.js button titles and refactor components
